user-management-service/requirements.txt
anyio==3.7.0
bcrypt==4.0.1
certifi==2023.5.7
click==8.1.3
databases==0.7.0
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.97.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
loguru==0.7.0
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==23.1
passlib==1.7.4
pluggy==1.0.0
psycopg2-binary==2.9.6
pydantic==1.10.9
Pygments==2.15.1
PyJWT==2.8.0
pytest==7.3.2
pytest-mock==3.11.1
rich==13.4.2
sniffio==1.3.0
SQLAlchemy==1.4.48
starlette==0.27.0
typing_extensions==4.6.3
uvicorn==0.22.0
